不可删除非本人新建的记录

+++ src/main/java/com/song/archives/controller/AnliController.java
@@ -200,6 +200,14 @@ public class AnliController {
                 for (Long id:ids){
                     AnliInfoEntity entity = anliRepository.findOne(id);
                     operationLogInfo += entity.getTitle() + ",";
+
+                    if (!entity.getCreator().equals(getUser().getUsername())){
+                        msg = "不能删除非本人新建的记录";
+                        result.put("msg", msg);
+                        LoggerUtils.setLoggerFailed(request,msg);
+                        return result.toString();
+                    }
+
                     anliRepository.delete(entity);
 
                 }
